% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isocodes.R
\name{get_isocodes}
\alias{get_isocodes}
\title{Get ISO3 code of all World countries}
\usage{
get_isocodes()
}
\value{
A data frame with 3 columns:
  - region, the regions of the world
  - country, the countries of the world
  - iso3, the iso3 code of countries
}
\description{
Get the ISO3 code of all World countries from the Protected Planet website (based on the UNEP classification).
}
\details{
A Internet connexion is required.
}
\examples{
x <- get_isocodes()

head(x)

x[x[ , "region"] == "Africa", ]
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}
}
