% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regions.R
\name{get_regions}
\alias{get_regions}
\title{Get World regions names}
\usage{
get_regions()
}
\value{
A vector of World regions names.
}
\description{
Get World regions names from the Protected Planet website (based on the UNEP classification).
}
\details{
A Internet connexion is required.
}
\examples{
get_regions()
}
\seealso{
\link{\code{get_isocodes()}}, \link{\code{get_countries()}}, \link{\code{get_shapefile()}}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}
}
