% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Check Protected Planet API Token}
\usage{
get_token(key = "WDPA_KEY")
}
\arguments{
\item{key}{a character providing the WDPA token name stored in the
\code{.Renviron} file. Default is \code{WDPA_KEY}.}
}
\value{
A vector of length one with the value of the API token.
}
\description{
This function checks if the user has stored a valid Protected Planet API
token in the R environment file (\code{.Renviron}).
}
\details{
Before using this package for the first time, the user must follow
these steps:
\itemize{
\item Fill in the form available at:
\url{https://api.protectedplanet.net/request} to obtain a personal
API token;
\item Store the token in the \code{.Renviron} file under the key \code{WDPA_KEY}.
User can use the function \code{usethis::edit_r_environ} to open this
file;
\item Restart R.
}
}
\examples{
\dontrun{
## Check if Protected Planet API token is stored ----
worldpa::get_token()
}
}
