% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wdpa.R
\name{get_wdpa}
\alias{get_wdpa}
\title{Download World Protected Areas Shapefile}
\usage{
get_wdpa(regions, countries, path = ".")
}
\arguments{
\item{regions}{[string] A vector of world regions (see \code{get_regions()} for the correct spelling).}

\item{countries}{[string] A vector of world countries (see \code{get_countries()} for the correct spelling).}

\item{path}{[string] The path (directory) to write downloaded files.}
}
\value{
Nothing.
}
\description{
Download World protected areas shapefiles from the Protected Planet website.
}
\details{
A Internet connexion is required.
}
\examples{
\dontrun{
# Download protected area for the world
get_wdpa()
}

\dontrun{
# Download protected area for the African continent
get_regions()
get_wdpa(regions = "Africa")
}

# Download protected area for Algeria
get_countries()
get_wdpa(countries = "Algeria")

\dontrun{
# Download protected area for Spain and Portugal
get_countries()
get_wdpa(countries = c("Spain", "Portugal")
}
}
\seealso{
\code{get_isocodes}, \code{get_countries}, \code{get_regions}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}
}
