% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldpa-package.R
\docType{package}
\name{worldpa-package}
\alias{worldpa}
\alias{worldpa-package}
\title{worldpa: An Interface to the World Database on Protected Areas (WDPA)}
\description{
This package is an interface to the World Database on Protected
    Areas <https://www.protectedplanet.net> and its API
    <https://api.protectedplanet.net>. User can download terrestrial and marine
    protected areas for the world countries (one country at the time).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/worldpa}
  \item Report bugs at \url{https://github.com/frbcesab/worldpa/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID})

Other contributors:
\itemize{
  \item Kevin Cazelles \email{kevin.cazelles@gmail.com} (\href{https://orcid.org/0000-0001-6619-9874}{ORCID}) [contributor, reviewer]
}

}
\keyword{internal}
