% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countries.R
\name{get_countries}
\alias{get_countries}
\title{Get World Countries Listed by the UNEP-WCMC}
\usage{
get_countries(sleep = 0, key = "WDPA_KEY")
}
\arguments{
\item{sleep}{a numeric specifying the time interval (in seconds) to suspend
between each API request.}

\item{key}{a character providing the WDPA token name stored in the
\code{.Renviron} file. Default is \code{WDPA_KEY}.}
}
\value{
A data frame with the following information (columns) for each
World countries (rows):
\describe{
\item{region_name}{the name of the region}
\item{region_iso2}{the ISO-2 code of the region}
\item{country_name}{the name of the country}
\item{country_iso3}{the ISO-3 code of the country}
\item{pas_count}{the number of protected areas per country}
}
}
\description{
This function gets the list of world countries listed by the UNEP-WCMC using
the Protected Planet API \url{https://api.protectedplanet.net/}.
}
\examples{
\dontrun{
## List World Countries ----
countries <- worldpa::get_countries(sleep = 0.25)
head(countries)

##   region_name region_iso2 country_name country_iso3 pas_count
## 1      Africa          AF      Algeria          DZA        78
## 2      Africa          AF       Angola          AGO        14
## 3      Africa          AF        Benin          BEN        64
## 4      Africa          AF     Botswana          BWA        22
## 5      Africa          AF Burkina Faso          BFA       112
## 6      Africa          AF      Burundi          BDI        21
}
}
\seealso{
\code{wdpa_countries}, \code{get_wdpa}
}
