% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countries.R
\name{get_countries}
\alias{get_countries}
\title{Get World countries names}
\usage{
get_countries()
}
\value{
A vector of World countries names.
}
\description{
Get World countries names from the Protected Planet website (based on the UNEP classification).
}
\details{
A Internet connexion is required.
}
\examples{
get_countries()
}
\seealso{
\code{get_isocodes}, \code{get_regions}, \code{get_wdpa}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}
}
