% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countries.R
\name{get_countries}
\alias{get_countries}
\title{Get the list of countries listed by the UNEP}
\usage{
get_countries(sleep = 0)
}
\arguments{
\item{sleep}{The time interval to suspend between each API request (in seconds).}
}
\value{
A data frame with countries informations. See `wdpa_countries` for further details.
}
\description{
This function gets the list of countries listed by the UNEP using the WDPA API.
}
\examples{

## get_countries(sleep = 0.25)
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}
}
