% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Check Protected Planet API token}
\usage{
get_token(key = "WDPA_KEY")
}
\arguments{
\item{key}{[string] The name of the token value stored in the .Renviron file (recommended key name: 'WDPA_KEY').}
}
\value{
A vector of length one with the value of the API token.
}
\description{
This function checks if the user has stored a valid Protected Planet API token in the R environment (.Renviron) file under the key 'WDPA_KEY'.
}
\details{
Before using this package for the first time, the user must follow these steps:
\itemize{
  \item Fill in the form available at: \link{https://api.protectedplanet.net/request} to obtain a personal API token;
  \item Store the token in the .Renviron file under the key 'WDPA'. User can use the function \code{usethis::edit_r_environ()}.
}
}
\examples{

## get_token()
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}
}
