% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wdpa.R
\name{get_wdpa}
\alias{get_wdpa}
\title{Download Protected Areas by Country}
\usage{
get_wdpa(isocode)
}
\arguments{
\item{isocode}{The ISO-3 code of the country.}
}
\value{
A MULTIPOLYGON Simple feature of protected areas defined in the EPSG 4326. The shapefile is also written on the hard drive (in the current directory).
}
\description{
This function downloads protected areas for one country using the WDPA API.
}
\examples{
## Get Protected Areas for Georgia

## data(wdpa_countries)

## wdpa_countries[grep("Geor", wdpa_countries[ , "country_name"]), ]
##     region_name region_iso2                                 country_name country_iso3 pas_count
## 133      Europe          EU                                      Georgia          GEO        89
## 235       Polar          PO South Georgia and the South Sandwich Islands          SGS         1

## get_wdpa(isocode = "GEO")
## Simple feature collection with 89 features and 6 fields
## geometry type:  MULTIPOLYGON
## dimension:      XY
## bbox:           xmin: 40.31 ymin: 41.093 xmax: 46.736 ymax: 43.544
## epsg (SRID):    4326
## proj4string:    +proj=longlat +datum=WGS84 +no_defs
## First 10 features:
##    wdpa_id           pa_name country_iso3 is_marine           designation iucn_category
## 1     1652           Borjomi          GEO     FALSE Strict Nature Reserve            Ia
## 2     1653         Lagodekhi          GEO     FALSE Strict Nature Reserve            Ia
## 3     1654             Ritsa          GEO     FALSE Strict Nature Reserve            Ia
## 4     1656         Kintrishi          GEO     FALSE Strict Nature Reserve            Ia
## 5     1657           Liakhvi          GEO     FALSE Strict Nature Reserve            Ia
## 6     1660        Vashlovani          GEO     FALSE Strict Nature Reserve            Ia
## 7     1664 Bichvinta-Miusera          GEO     FALSE Strict Nature Reserve            Ia
## 8     1665       Mariamjvari          GEO     FALSE Strict Nature Reserve            Ia
## 9     1666          Kolkheti          GEO      TRUE         National Park            II
## 10    1667          Sataplia          GEO     FALSE Strict Nature Reserve            Ia
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}
}
