% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wdpa.R
\name{get_wdpa}
\alias{get_wdpa}
\title{Download Protected Areas by Country}
\usage{
get_wdpa(isocode, key = "WDPA_KEY")
}
\arguments{
\item{isocode}{a character specifying the ISO-3 code of the country.}

\item{key}{a character providing the WDPA token name stored in the
\code{.Renviron} file. Default is \code{WDPA_KEY}.}
}
\value{
A \code{MULTIPOLYGON} Simple Feature of protected areas defined in the
EPSG 4326 with the following 19 attributes:
\describe{
\item{wdpa_id}{the ID of the protected area on the WDPA database}
\item{pa_name}{the name of the protected area on the WDPA database}
\item{original_name}{the original name of the protected area}
\item{country_iso3}{the ISO-3 code of the country}
\item{country}{the name of the country}
\item{owner_type}{the owner of the protected area}
\item{is_marine}{a boolean: \code{TRUE} for marine and \code{FALSE} for terrestrial
protected area}
\item{designation}{the designation of the protected area}
\item{iucn_category}{the IUCN category}
\item{no_take_status}{the no take status}
\item{reported_area}{the (reported) area of the protected area}
\item{reported_marine_area}{the (reported) area of the marine part of the
protected area}
\item{no_take_area}{the (reported) area of the no take part of the
protected area}
\item{legal_status}{the legal status of the protected area}
\item{management_authority}{the management authority}
\item{governance}{the type of gouvernance}
\item{management_plan}{the URL of the management plan}
\item{protected_planet}{the URL of the protected area page on the Protected
Planet website}
\item{legal_status_updated}{the date of the last update of the legal status}
}

Spatial shapefiles (ESRI format) are also written on the hard drive in a new
folder \strong{XXX_protectedareas/} (where \code{XXX} is the ISO-3 code of the country).
}
\description{
This function downloads protected areas for one country using the WDPA API.
}
\examples{
\dontrun{
## Get the ISO-3 code of Georgia ----
data("wdpa_countries", package = "worldpa")
wdpa_countries[grep("^Georgia", wdpa_countries$"country_name"), 3:4]

##                                     country_name country_iso3
## 133                                      Georgia          GEO
## 235 South Georgia and the South Sandwich Islands          SGS

## Download Protected Areas for Georgia ----
worldpa::get_wdpa(isocode = "GEO")
}
}
\seealso{
\code{wdpa_countries}, \code{get_countries}
}
