% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wdpa.R
\name{get_wdpa}
\alias{get_wdpa}
\title{Download World Protected Areas Shapefile}
\usage{
get_wdpa(regions, countries, path = ".")
}
\arguments{
\item{regions}{[string] A vector of world regions (see \code{get_regions()} for the correct spelling).}

\item{countries}{[string] A vector of world countries (see \code{get_countries()} for the correct spelling).}

\item{path}{[string] The path (directory) to write downloaded files.}
}
\value{
Nothing.
}
\description{
Download World protected areas shapefiles from the Protected Planet website.
}
\details{
A Internet connexion is required.
}
\examples{
# vignette(topic = "worldpa")
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}
}
