% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldpa-package.R
\docType{package}
\name{worldpa-package}
\alias{worldpa}
\alias{worldpa-package}
\title{worldpa: API-based R package to download World protected areas shapefile}
\description{
API-based R package to download World protected areas shapefiles from the Protected Planet website (https://www.protectedplanet.net).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/worldpa}
  \item Report bugs at \url{https://github.com/frbcesab/worldpa/issues}
}

}
\keyword{internal}
